refactor: Refactored training scripts
